+++ internal/server/web/static/js/restrictions.js
@@ -1,13 +1,47 @@
-const $editorMode = $('#restr-editor-mode');
+function editorMode(prefix = "") {
+    return $('#' + prefix + 'restr-editor-mode');
+}
 
-function updateRestrIcons() {
+function restrictionsArea(prefix = "") {
+    return $('#' + prefix + 'restrictionsArea');
+}
+
+function restrIconTime(prefix = "") {
+    return $('#' + prefix + 'r-icon-time');
+}
+
+function restrIconIP(prefix = "") {
+    return $('#' + prefix + 'r-icon-ip');
+}
+
+function restrIconScope(prefix = "") {
+    return $('#' + prefix + 'r-icon-scope');
+}
+
+function restrIconAud(prefix = "") {
+    return $('#' + prefix + 'r-icon-aud');
+}
+
+function restrIconUsages(prefix = "") {
+    return $('#' + prefix + 'r-icon-usages');
+}
+
+function getRestrictionsData(prefix = "") {
+    return getPrefixData(prefix)['restrictions']['restrictions'];
+}
+
+function setRestrictionsData(data, prefix = "") {
+    getPrefixData(prefix)['restrictions']['restrictions'] = data;
+}
+
+function updateRestrIcons(prefix = "") {
     let howManyClausesRestrictIP = 0;
     let howManyClausesRestrictScope = 0;
     let howManyClausesRestrictAud = 0;
     let howManyClausesRestrictUsages = 0;
     let expires = 0;
     let doesNotExpire = false;
-    restrictions.forEach(function (r) {
+    getRestrictionsData(prefix).forEach(function (r) {
         if (r['scope'] !== undefined) {
             howManyClausesRestrictScope++;
         }
@@ -36,70 +70,67 @@ function updateRestrIcons() {
     if (doesNotExpire) {
         expires = 0;
     }
-    let iconTime = $('#r-icon-time');
-    let iconIP = $('#r-icon-ip');
-    let iconScope = $('#r-icon-scope');
-    let iconAud = $('#r-icon-aud');
-    let iconUsages = $('#r-icon-usages');
-    if (howManyClausesRestrictIP === restrictions.length && restrictions.length > 0) {
-        iconIP.addClass('text-success');
-        iconIP.removeClass('text-warning');
-        iconIP.removeClass('text-danger');
-        iconIP.attr('data-original-title', "The IPs from which this token can be used are restricted.");
+    let restr = getRestrictionsData(prefix);
+    if (howManyClausesRestrictIP === restr.length && restr.length > 0) {
+        restrIconIP(prefix).addClass('text-success');
+        restrIconIP(prefix).removeClass('text-warning');
+        restrIconIP(prefix).removeClass('text-danger');
+        restrIconIP(prefix).attr('data-original-title', "The IPs from which this token can be used are restricted.");
     } else {
-        iconIP.addClass('text-warning');
-        iconIP.removeClass('text-success');
-        iconIP.removeClass('text-danger');
-        iconIP.attr('data-original-title', "This token can be used from any IP.");
+        restrIconIP(prefix).addClass('text-warning');
+        restrIconIP(prefix).removeClass('text-success');
+        restrIconIP(prefix).removeClass('text-danger');
+        restrIconIP(prefix).attr('data-original-title', "This token can be used from any IP.");
     }
-    if (howManyClausesRestrictScope === restrictions.length && restrictions.length > 0) {
-        iconScope.addClass('text-success');
-        iconScope.removeClass('text-warning');
-        iconScope.removeClass('text-danger');
-        iconScope.attr('data-original-title', "This token has restrictions for scopes.");
+    if (howManyClausesRestrictScope === restr.length && restr.length > 0) {
+        restrIconScope(prefix).addClass('text-success');
+        restrIconScope(prefix).removeClass('text-warning');
+        restrIconScope(prefix).removeClass('text-danger');
+        restrIconScope(prefix).attr('data-original-title', "This token has restrictions for scopes.");
     } else {
-        iconScope.addClass('text-warning');
-        iconScope.removeClass('text-success');
-        iconScope.removeClass('text-danger');
-        iconScope.attr('data-original-title', "This token can use all configured scopes.");
+        restrIconScope(prefix).addClass('text-warning');
+        restrIconScope(prefix).removeClass('text-success');
+        restrIconScope(prefix).removeClass('text-danger');
+        restrIconScope(prefix).attr('data-original-title', "This token can use all configured scopes.");
     }
-    if (howManyClausesRestrictAud === restrictions.length && restrictions.length > 0) {
-        iconAud.addClass('text-success');
-        iconAud.removeClass('text-warning');
-        iconAud.removeClass('text-danger');
-        iconAud.attr('data-original-title', "This token can only obtain access tokens with restricted audiences.");
+    if (howManyClausesRestrictAud === restr.length && restr.length > 0) {
+        restrIconAud(prefix).addClass('text-success');
+        restrIconAud(prefix).removeClass('text-warning');
+        restrIconAud(prefix).removeClass('text-danger');
+        restrIconAud(prefix).attr('data-original-title', "This token can only obtain access tokens with restricted" +
+            " audiences.");
     } else {
-        iconAud.addClass('text-warning');
-        iconAud.removeClass('text-success');
-        iconAud.removeClass('text-danger');
-        iconAud.attr('data-original-title', "This token can obtain access tokens with any audiences.");
+        restrIconAud(prefix).addClass('text-warning');
+        restrIconAud(prefix).removeClass('text-success');
+        restrIconAud(prefix).removeClass('text-danger');
+        restrIconAud(prefix).attr('data-original-title', "This token can obtain access tokens with any audiences.");
     }
-    if (howManyClausesRestrictUsages === restrictions.length && restrictions.length > 0) {
-        iconUsages.addClass('text-success');
-        iconUsages.removeClass('text-warning');
-        iconUsages.removeClass('text-danger');
-        iconUsages.attr('data-original-title', "This token can only be used a limited number of times.");
+    if (howManyClausesRestrictUsages === restr.length && restr.length > 0) {
+        restrIconUsages(prefix).addClass('text-success');
+        restrIconUsages(prefix).removeClass('text-warning');
+        restrIconUsages(prefix).removeClass('text-danger');
+        restrIconUsages(prefix).attr('data-original-title', "This token can only be used a limited number of times.");
     } else {
-        iconUsages.addClass('text-warning');
-        iconUsages.removeClass('text-success');
-        iconUsages.removeClass('text-danger');
-        iconUsages.attr('data-original-title', "This token can be used an infinite number of times.");
+        restrIconUsages(prefix).addClass('text-warning');
+        restrIconUsages(prefix).removeClass('text-success');
+        restrIconUsages(prefix).removeClass('text-danger');
+        restrIconUsages(prefix).attr('data-original-title', "This token can be used an infinite number of times.");
     }
     if (expires === 0) {
-        iconTime.addClass('text-danger');
-        iconTime.removeClass('text-success');
-        iconTime.removeClass('text-warning');
-        iconTime.attr('data-original-title', "This token does not expire!");
+        restrIconTime(prefix).addClass('text-danger');
+        restrIconTime(prefix).removeClass('text-success');
+        restrIconTime(prefix).removeClass('text-warning');
+        restrIconTime(prefix).attr('data-original-title', "This token does not expire!");
     } else if ((expires - Date.now() / 1000) > 3 * 24 * 3600) {
-        iconTime.addClass('text-warning');
-        iconTime.removeClass('text-success');
-        iconTime.removeClass('text-danger');
-        iconTime.attr('data-original-title', "This token is long-lived.");
+        restrIconTime(prefix).addClass('text-warning');
+        restrIconTime(prefix).removeClass('text-success');
+        restrIconTime(prefix).removeClass('text-danger');
+        restrIconTime(prefix).attr('data-original-title', "This token is long-lived.");
     } else {
-        iconTime.addClass('text-success');
-        iconTime.removeClass('text-warning');
-        iconTime.removeClass('text-danger');
-        iconTime.attr('data-original-title', "This token expires within 3 days.");
+        restrIconTime(prefix).addClass('text-success');
+        restrIconTime(prefix).removeClass('text-warning');
+        restrIconTime(prefix).removeClass('text-danger');
+        restrIconTime(prefix).attr('data-original-title', "This token expires within 3 days.");
     }
 }
 
@@ -116,46 +147,48 @@ function newJSONEditor(textareaID) {
     });
 }
 
-function fillJSONEditor() {
-    $('#restrictionsArea').val(JSON.stringify(restrictions, null, 4));
+function fillJSONEditor(prefix = "") {
+    restrictionsArea(prefix).val(JSON.stringify(getRestrictionsData(prefix), null, 4));
 }
 
-function updateRestrFromJSONEditor() {
+function updateRestrFromJSONEditor(prefix) {
     let r = [];
-    let res = $('#restrictionsArea');
     try {
-        r = JSON.parse(res.val());
-        res.removeClass('is-invalid');
-        res.addClass('is-valid');
+        r = JSON.parse(restrictionsArea(prefix).val());
+        restrictionsArea(prefix).removeClass('is-invalid');
+        restrictionsArea(prefix).addClass('is-valid');
     } catch (e) {
-        res.removeClass('is-valid');
-        res.addClass('is-invalid');
+        restrictionsArea(prefix).removeClass('is-valid');
+        restrictionsArea(prefix).addClass('is-invalid');
         return;
     }
-    restrictions = r;
-    updateRestrIcons();
+    setRestrictionsData(r, prefix);
+    updateRestrIcons(prefix);
 }
 
-updateRestrIcons();
-newJSONEditor('restrictionsArea');
 
-$(function () {
-    $editorMode.prop("checked", true);
-    $editorMode.on('change', function () {
-        let htmlEdit = $('#restr-easy-editor');
-        let jsonEdit = $('#restr-json-editor');
+function initRestr(prefix = "", ...next) {
+    newJSONEditor(prefix + 'restrictionsArea');
+    editorMode(prefix).bootstrapToggle("on");
+    editorMode(prefix).on('change', function () {
+        let htmlEdit = $('#' + prefix + 'restr-easy-editor');
+        let jsonEdit = $('#' + prefix + 'restr-json-editor');
         if ($(this).prop("checked")) { // easy editor
-            RestrToGUI();
+            RestrToGUI(prefix);
             jsonEdit.hideB();
             htmlEdit.showB();
         } else { // JSON editor
-            fillJSONEditor();
+            fillJSONEditor(prefix);
             jsonEdit.showB();
             htmlEdit.hideB();
         }
     });
-})
+    updateRestrIcons(prefix);
+    initRestrGUI(prefix);
+    doNext(...next);
+}
 
-$('#restr-expand').on('click', function () {
-    $('#restr-editor-wrap').toggleClass('d-none');
+$('.restr-expand').on('click', function () {
+    let prefix = this.getAttribute("instance-prefix");
+    $('#' + prefix + 'restr-editor-wrap').toggleClass('d-none');
 });